tweak Dockerfiles, mostly for better caching
These minor edits to the Dockerfiles tidy up the files and make better
use of docker layers for caching. The best change is moving the
`ARG THELOUNGE_VERSION` along with the npm install to then end of the
Dockerfile. Bumping the version needlessy busts the cache for everything
that comes after. It is safe to move the install to the very end since
everything else is either docker metadata or runtime configuration that
does not affect installation.

The VOLUME command will create the directory if it doesn't exist,
so we can remove the mkdir.

At runtime, the ENTRYPOINT is always run before CMD or container args,
having the Dockerfile reflect that just seems nice.

I moved the COPY docker-entrypoint.sh before lounge gets installed
because it changes less often than lounge version bumps so we get to
make use of the cache more often.

I moved the version ARG until the very end so that we don't bust the
cache unnecessarily for stages that are not affected by the version.

I also rearranged the EXPOSE and install stages because install does not
depend on it and it lets us make better use of the cache.