Build for ARM (ARM64 + ARMv7)
With this commit we switch from "Automated Builds"[1] to using a GitHub
Actions for building the Docker images. This makes it possible to use
buildx[2] and build the images for other platforms besides amd64.

The workflow is only triggered when a tag is pushed and the latest tag
(`latest` or `alpine`) is only pushed when the tag match the following
regex: "^[0-9]*\.[0-9]*\.[0-9]*$".

[1] https://docs.docker.com/docker-hub/builds/
[2] https://github.com/docker/buildx

Fix #99